requirements.txt
# --- [API & Framework] ---
fastapi==0.129.0
uvicorn[standard]==0.41.0
python-dotenv==1.2.1
python-dateutil==2.9.0.post0
python-multipart==0.0.22
pydantic==2.12.5
pydantic_settings==2.13.0

# --- [AI Core]
scikit-learn==1.6.1
torch==2.10.0
transformers==5.2.0
joblib==1.5.3
numpy==2.4.2
pandas==3.0.1

# --- [Scraping & Data] ---
beautifulsoup4==4.14.3
selenium==4.40.0
selenium-wire==5.1.0
undetected-chromedriver==3.5.5
webdriver-manager==4.0.2

# --- [DB & Infrastructure] ---
SQLAlchemy==2.0.46
redis==7.2.0
google-generativeai==0.8.6

# --- [Security & Utils] ---
python-jose==3.5.0
passlib[bcrypt]==1.7.4
Requests==2.32.5
protobuf==5.29.6
certifi==2026.1.4